a few things
-Add customization for delta number color
-Replace lunar ability exclusion with a void survivor ability inclusion. Because mods or DLC could add either new void survivor abilities or new items which replace the special ability, an inclusion works better.
-replace hardcoded 25% with values gotten from the ability asset. addon to the above, since they share the same dictionary. same reasoning, values could change, mods, etc.

The whole code for the deltas needs work. because we can grab the values themselves, the code can be consolidated into one "meter delta generation" function. This also allows more flexibility in case a mod or DLC adds an ability which adds corruption when not corrupted, or removes it when corrupted.